Correctly wrap System::session into CFReleaser in macOS
